+++ include/BLIB/Interfaces/GUI/Renderers/DefaultRenderer.hpp
@@ -48,7 +48,7 @@ public:
 
     /// @see Renderer::renderComboBoxDropdownBoxes
     virtual void renderComboBoxDropdownBoxes(sf::RenderTarget& target, sf::RenderStates states,
-                                             const ComboBox& box, const sf::Vector2i& optionSize,
+                                             const ComboBox& box, const sf::Vector2f& optionSize,
                                              unsigned int optionCount,
                                              unsigned int mousedOption) const override;
