+++ include/BLIB/Interfaces/GUI/Renderers/Renderer.hpp
@@ -92,7 +92,7 @@ public:
      *                     over. None are moused over if mousedOption >= optioncount
      */
     virtual void renderComboBoxDropdownBoxes(sf::RenderTarget& target, sf::RenderStates states,
-                                             const ComboBox& box, const sf::Vector2i& optionSize,
+                                             const ComboBox& box, const sf::Vector2f& optionSize,
                                              unsigned int optionCount,
                                              unsigned int mousedOption) const = 0;
